Refactored components into subdirectory
Refactored JS and style components into a subdirectory, added
react-bootstrap dependency
