+++ package.json
@@ -4,12 +4,14 @@
   "private": true,
   "dependencies": {
     "abcjs": "^5.6.4",
+    "bootstrap": "^4.3.1",
     "history": "^4.7.2",
     "markdown-it": "^8.4.2",
     "markdown-it-music": "github:music-markdown/markdown-it-music",
     "node-sass": "^4.11.0",
     "query-string": "^6.2.0",
     "react": "^16.8.1",
+    "react-bootstrap": "^1.0.0-beta.5",
     "react-dom": "^16.8.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "^2.1.5"
