+++ src/components/App.scss
@@ -1,4 +1,4 @@
-@import "../node_modules/markdown-it-music/renderers/renderer.css";
+@import "../../node_modules/markdown-it-music/renderers/renderer.css";
 
 .App {
 }
