+++ src/index.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import App from './components/App';
 import registerServiceWorker from './registerServiceWorker';
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
 ReactDOM.render(
-    <Router>
-        <Route path='/' exact component={App} />
-    </Router>,
-    document.getElementById('root'));
+  <Router>
+    <Route path='/' exact component={App} />
+  </Router>,
+  document.getElementById('root'));
 registerServiceWorker();
